services/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pydantic==2.8.2
qdrant-client==1.10.1
sentence-transformers==3.0.1
numpy==1.26.4
scikit-learn==1.4.2
google-generativeai==0.3.2
transformers==4.37.2
torch>=2.2.0
python-dotenv==1.0.0
aiohttp==3.9.1
psutil==5.9.8
redis==5.0.1
slowapi==0.1.9
langgraph==0.2.16
langchain==0.2.16
